+++ utbot-analytics/src/main/kotlin/org/utbot/predictors/LinearStateRewardPredictor.kt
@@ -1,11 +1,15 @@
 package org.utbot.predictors
 
+import mu.KotlinLogging
 import org.utbot.analytics.UtBotAbstractPredictor
+import org.utbot.framework.PathSelectorType
 import org.utbot.framework.UtSettings
 import smile.math.matrix.Matrix
 import java.io.File
 
 private const val DEFAULT_WEIGHT_PATH = "linear.txt"
+
+private val logger = KotlinLogging.logger {}
 
 /**
  * Last weight is bias
@@ -24,7 +28,18 @@ private fun loadWeights(path: String): Matrix {
 
 class LinearStateRewardPredictor(weightsPath: String = DEFAULT_WEIGHT_PATH) :
     UtBotAbstractPredictor<List<List<Double>>, List<Double>> {
-    private val weights = loadWeights(weightsPath)
+    private lateinit var weights: Matrix
+
+    init {
+        try {
+            weights = loadWeights(weightsPath)
+        } catch (e: Exception) {
+            logger.error(e) {
+                "Error while initialization of LinearStateRewardPredictor. Changing pathSelectorType on INHERITORS_SELECTOR"
+            }
+            UtSettings.pathSelectorType = PathSelectorType.INHERITORS_SELECTOR
+        }
+    }
 
     override fun predict(input: List<List<Double>>): List<Double> {
         // add 1 to each feature vector
