+++ utbot-junit-contest/src/main/kotlin/org/utbot/contest/ContestEstimator.kt
@@ -2,7 +2,6 @@ package org.utbot.contest
 
 import mu.KotlinLogging
 import org.utbot.analytics.EngineAnalyticsContext
-import org.utbot.analytics.Predictors
 import org.utbot.common.FileUtil
 import org.utbot.common.bracket
 import org.utbot.common.info
@@ -22,7 +21,6 @@ import org.utbot.framework.plugin.api.util.UtContext
 import org.utbot.framework.plugin.api.util.id
 import org.utbot.framework.plugin.api.util.withUtContext
 import org.utbot.instrumentation.ConcreteExecutor
-import org.utbot.predictors.NNStateRewardPredictorSmile
 import java.io.File
 import java.io.FileInputStream
 import java.net.URLClassLoader
